JIRA:VR-737 use size_t for NStr::Find result
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@79040 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/app/asnval/asnval.cpp
@@ -446,7 +446,7 @@ void CAsnvalApp::ValidateOneFile(const string& fname)
             }
         } catch (CException &e) {
             string errstr = e.what();
-            int pos = NStr::Find (errstr, "duplicate Bioseq id");
+            size_t pos = NStr::Find (errstr, "duplicate Bioseq id");
             if (pos != NPOS && m_verbosity == eVerbosity_XML) {
                 CRef<CValidError> eval(new CValidError());
                 errstr.erase(0,pos);
